Added logic to detect and alert of potentially stuck deployment jobs (#162)
## Summary:
We are having random deploy jobs getting stuck causing all sorts of nasty issues in the deploy queue with seeming stuck jobs after a user has abort their deployment. The theory is that abort signals from jenkins to threads are not working because the threads are in uninterruptible cycles.  To help us determine this I have added slack alerts to the `#infrastructure-deploys` channel to let us know that we pontentially have deploy-webapp jobs that are stuck.

I have added logic to trigger an abort but I have it disabled to start with to ensure that we have the correct logic in place before potentially start to abort jobs that shouldn't be.  Once this is out and we can test it, we can enable the logic to abort build that we deem "stuck".

Issue: https://khanacademy.atlassian.net/browse/INFRA-9495

## Test plan:
Monitor the deploy and verify that we don't have deploy jobs suddenly aborting.  Icing on the cake, try to reproduce the stuck aborted jobs to test the slack logging logic.

1. queue a deploy.
2. when asked to `set default` by sun, monitor your deploy-webapp job and try to abort when we are making the request to buildmaster to get the `prompt-status`.
3. If the job is stuck, then we will keep looping asking for the prompt-status.
4. Check #infrastucture-deploys channel for informational message about the deploy.

This is pretty tricky to reproduce... And best done when nobody is in the queue.

Author: MiguelCastillo

Reviewers: MiguelCastillo, dnerdy, csilvers, #devops

Required Reviewers:

Approved By: csilvers

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/162