[🔥AUDIT🔥] Print class name for exceptions in notify.log (#165)
🖍 _This is an audit!_ 🖍

## Summary:
We have a theory that the call to gcloud is throwing an exception in notify.log when jenkins is aborting a job, but the exception message is null. And there is no clear way to know what exception was thrown. So I am including the class name for the exception being thrown hoping to help us verify the theory.

Issue: https://khanacademy.atlassian.net/browse/INFRA-9495

## Test plan:
This one is hard to reproduce, so hoping we can see the message when a job is stuck which we are reporting in `#infrastructure-deploys`

Author: MiguelCastillo

Auditors: #devops, csilvers

Required Reviewers:

Approved By:

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/165