Make sure we check for abort every time we catch an exception. (#164)
## Summary:
If a jenkins job is aborted, every process gets passed an
AbortException.  But it's possible to swallow that exception if your
code is in a try/catch.

This PR looks at every `catch` in our codebase, and if the `catch`
doesn't always rethrow anyway, I add a call to a function that checks
if the job has been aborted, and if so rethrows the exception (which
is presumably AbortException).

That should resolve issues like
https://jenkins.khanacademy.org/job/deploy/job/deploy-webapp/16256
where a job got aborted but stuff kept on running anyway, because the
`notify.log()` function swallowed the abort exception.

Issue: https://khanacademy.atlassian.net/browse/INFRA-9495

## Test plan:
Am mostly just concerned this doesn't break anything; we'll have to
land it and see.

Author: csilvers

Reviewers: MiguelCastillo, #devops, dnerdy

Required Reviewers:

Approved By: MiguelCastillo

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/164