Updating owner to infra-platform for code reviewing purposes
We have transitioned ownership of this code to infra-platform so Im reflecting that in the github code owners so that the correct group of folks get code reviews.

Issue: https://khanacademy.slack.com/archives/C01120CNCS0/p1698247174759139

Test plan:
Wait for a PR - such as this one - and verify the correct group is added.