Make sure jenkins-server runs as prod-deploy@.
I had activated a service-account in #169, but I didn't realize that
this command makes that service-account the default for all gcloud
operations.  We still want the default to be done as
prod-deploy@khanacademy.org!  And only use the service-account with
the `--account` flag.  This new command does that.

Issue: https://khanacademy.slack.com/archives/C096UP7D0/p1698352831330619

Test plan:
Ran the command manually on jenkins-server with success.