Redo how we track failed jobs in weekly-maintenace.sh.
The backup_network job was failing, but the script wasn't reporting
that it failed.  This is because a subshell exits with the result of
the lsat command, which in this case is an `echo` which always
suceeds.  Fix that.

Issue: https://jenkins.khanacademy.org/job/misc/job/webapp-maintenance/268/execution/node/910/log/

Test plan:
Fingers crossed