Finally get the script to actually report failed jobs correctly.
`set -e` does not behave in a useful way, sadly.  There is no way
(that I found) to have `set -e` fail a function if any command in that
function fails, but then have the caller of the function continue to
run even if the function call fails.  (This is because a `||` or a
`&&` disables the `-e` behavior in all code on the left side of the
operator.)

The only way to get the behavior I want -- that each individual job
runs with `set -e` semantics, but if a job fails we just mark it and
continue to the next job -- is to run the jobs in a separate process.
I can set `-e` in the separate process, while the driver runs without
`-e` and just keeps track of failures itself.

It's arguably easier to follow along what's going on now, though I'm
sad to have to split the code into two files.

Issue: none

Test plan:
From the parent-dir of this repo, I ran:
```
jenkins-jobs/weekly-maintenance.sh clean_docker
```
with success.

I then changed `clean_publish_worker` to have `false` as its first
command, and ran:
```
jenkins-jobs/weekly-maintenance.sh clean_publish_worker clean_docker
```
and it exited with rc 1, saying:
```
THE FOLLOWING JOBS FAILED:  clean_publish_worker
```
but the clean-docker command still ran (as I could see from the output
of running the script).