Let the update-devserver-images job take longer.  A lot longer.
Now that we're cross-compiling all our Go binaries on arm, using
emulation, it can take like 20 minutes to build the containers for a
single service.  So let's give it 10 hours to finish totally.

(A better solution is to use cross-compilation instead of emulation,
but that is a task for another day.)

Issue: none

Test plan:
Fingers crossed