Add more precise abort check
I added some logic in https://github.com/Khan/jenkins-jobs/pull/181 to send a message to slack if a job was aborted while in a call to `sh` and `sh` never throws an exception that aborts the jenkins job.  In this follow up change I am making the check more precise to ensure the job is in the `ABORTED` state.

Issue: https://khanacademy.atlassian.net/browse/INFRA-9495

Test plan:
groovy vars/notify.groovy