Add support for enabling the python3 virtualenv when running jobs. (#196)
## Summary:
We are porting our deploy and testing scripts from python2 to python3.
Some of them need third-party libraries such as `requests` or
`graphql`, which we are supporting via a python3 virtualenv that is
constructed in webapp's deploy/Makefile.

This PR adds Jenkins support to use that virtualenv.  We don't have a
user-friendly way to enable the virtualenv -- we don't expect users to
run these scripts, just Jenkins -- but the new
`withVirtualenv.python3()` does all the needful.

I had a choice: I could set up this new virtualenv in onWorker() and
onMaster(), like we do with the python2 virtualenv, or I could require
jobs to call it manually to set it up in the places where it's needed.
I chose to do the latter, since our goal is to replace all our python3
with Go (or js) over time, and I'm hoping that will allow us to remove
these calls as that happens, and eventually be left with no
withVirtualenv calls at all.  That is easier to do one at a time than
when it's globally done in onMaster/onWorker.

It does make it harder to know when to call the new python3() call,
since we need to make sure we're on the right machine when we do it.
For now, that's only an issue with webapp-test.groovy, which runs
across a lot of machines.

Issue: https://khanacademy.atlassian.net/browse/INFRA-9701

Test plan:

I ran `groovy` on all the files in this commit, with only the normal,
expected errors around `@CPS` and the like.

Author: csilvers

Reviewers: MiguelCastillo, csilvers

Required Reviewers:

Approved By: MiguelCastillo

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/196