Do not say "python" to run python binaries.
As we transition to python3, we can't be sure which version to run,
from the groovy file.  Let's let the shebang line on the script tell
us.

I tried to fix this in #202 but missed a case.

Issue: https://jenkins.khanacademy.org/job/deploy/job/build-webapp/46496/execution/node/425/log/

Test plan:
Fingers crossed