Reverting forcing make deps in deploy scripts
I forced the creation of the virtual env for python in the jobs to help me determine if that would make python happy creating the virtual env.  It does. But clearly that is not what we want to do.  So Im reverting while I figure out a sustainable solution.

Issue: https://khanacademy.slack.com/archives/C096UP7D0/p1707318372614619

Test plan:
Should be back to what we were before.  Im just undoing the change.