Do not try to deploy to fastly-staging in deploy-webapp.groovy.
That's too late!  This script doesn't get run until _after_ you say
`sun: set default`.  The whole point of the staging environment is to
be available _before_ you say `sun: set default`, when you're doing
the manual testing.

Sadly, doing this in build-webapp.groovy isn't right either, since
many of those get run in parallel, and there's only one fastly-staging
service.  If we really wanted to cram this logic in, we'd have to do
it in buildmaster, before showing the "time to set default" message to
the user.  But that's a really awkward place to put it.

We'll have to think more about what the right thing to do here.  The
whole fastly-staging thing isn't as useful as it might be, since it
doesn't work for first smoke tests, znd's, or any number of other
things we might want it to.  Perhaps fixing that problem will make it
obvious where this logic should move to (probably
build-webapp.groovy).

Issue: none

Test plan:
Noticing https://jenkins.khanacademy.org/job/deploy/job/deploy-webapp/20552/console