No longer create cloud-run tags on every deploy.
Now that we are routing to cloud-run in fastly (and Go), we no longer
need to create tags on every deploy to help us with the routing.  Huzzah!

This needs to wait to deploy until we are fully off of VCL and not
going back.  Since we did not backport the route-to-cloud-run logic to
VCL.

Issue: https://khanacademy.atlassian.net/browse/INFRA-10502

Test plan:
I ran `groovy jobs/build-webapp.groovy` without any new errors.