Get the build-current-sqlite script running again.
We turned this off when we moved to goliath, because the local
datastore only has user data in it, and that doesn't change very
often.  But it turns out it does change sometimes -- the permissions
granted to testadmin grow over time as more permissions get added to
our system (since testadmin is supposed to have all permissions).  So
there's some value in updating this data periodically.  This new job
does just that.

Issue: https://khanacademy.slack.com/archives/C04SEFXQBNU/p1740171086452009

Test plan:
I went to
   https://jenkins.khanacademy.org/job/misc/job/build-current-sqlite/1530/replay
and replaced the "Main Script" text with the new script text, except
for the addCronSchedule line, and ran it with success!