Post to `#whats-happening` when we deploy to fastly.
We have a goal of posting to the `#whats-happening` slack channel
whenever we deploy a change to our production system.  Currently,
changes to our fastly configs are noticed, via polling, by a cronjob
running on toby.  But now that all fastly deploys are (supposed to) go
through this jenkins job, it can do the reporting to #whats-happening
instead!

Once this is live, I'll disable the toby cronjob.  Or rather, I'll
repurpose it to notice when deploys to fastly are made via the UI and
not via the fastly deploy process.

Issue: https://khanacademy.slack.com/archives/C06T4JE5HCN/p1741644811156629

Test plan:
Fingers crossed, we'll see what happens next time we need to deploy
prod vcl!