Notify when someone manually deploys a change to fastly.
All of our fastly services are intended to be deployed via the
fastly-deploy jenkins job.  However, sometimes people deploy a fastly
service manually, by going to the fastly UI, hitting "clone", making
some changes, and hitting "save".  We do not want such changes to
happen, since there is no record of them in source control, and indeed
the change will be overwritten the next time a "proper" deploy
happens, via fastly-deploy.

This new script notices when such deploys happen.  It works in concert
with the fastly-deploy job to do its work: fastly-deploy says what
versions it has deployed, and we talk to the fastly API to find out
what versions are live that were *not* deployed by fastly-deploy.
When we see one, we alert in an alerting channel, and also in
production system.

A similar script used to exist on toby.  This PR moves it here, now
that it need to work in concert with the fastly-deploy jenkins job.
Once this is working I'll remove the old script from toby.

Issue: https://khanacademy.slack.com/archives/C06T4JE5HCN/p1741644811156629

Test plan:
I ran this manually on jenkins.