Add determine-webapp-services job (#306)
## Summary:
To simplify the logic for automated dequeues when doing an unattended
deploy, we should tell buildmaster what services a deploy should be
deploying before starting build-webapp. This is so failures during
build-webapp always have access to the list of services and therefore
the computed attendedness value.

Failures in webapp-test may need to trigger this job early in order to
tell if it should dequeue.

This job should not be started if the deployer has queued with the
SERVICES flag because that flag would override the result of this job.

This change will happen in 4 stages:

1. Deploy this PR to to perform testing with manual builds.
2. Update buildmaster to represent and call this job with a new state
   machine. At this time, build-webapp should overwrite the same
   services.
3. Update buildmaster to send the list of services received from this
   job to build-webapp so it doesn't need to recompute them.
4. Update build-webapp to no longer run webapp/deploy/should_deploy.py.

Issue: https://khanacademy.atlassian.net/browse/INFRA-10586

## Test plan:
1. Deploy jenkins job.
2. Manually start builds for the job.
3. Verify the list of services is output correctly into the logs.

Author: nathanjd

Reviewers: csilvers, nathanjd, dbraley, jwbron, MiguelCastillo

Required Reviewers:

Approved By: csilvers

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/306