fix(update-ownership-data,update-devserver-static-images): remove success messages from slack alerts to reduce slack noise
These success messages account for nearly 14% of the messages in #infrastructure for the last year. I think we only really care about when they fail and that they return to normal. Ive adjusted the when conditions to mimic the same behavior as "failureChannel" and removed failureChannel where it was redundant