[znd-safelist] Update the graphql safelist when we deploy znds
We are having an issue where updates to queries in from manually created znds are failing to pass our safelisting checks in the graphql gateway, which causes requests from znds to fail.  This regressed when we updated how we check for permissions of safelisted queries some time ago.

So in this PR I am making sure that we are adding queries from znds that engineers create themselves via jenkins. What Im not doing however, is priming queryplanner caches since these types of znds usually don't need to pre create all the queryplans as engineers tend to use a small subset of the UI.

I found this when I was testing adding the x-trace header to tasks we queue from our backends, which can be seen here https://cloudlogging.app.goo.gl/sHnpaBj8Fke8myCp7.

Issue: https://github.com/Khan/webapp/pull/31675#issuecomment-2828916672

Test plan:
Once these changes are done, deploy a znd with a changed opname and verify that works.