[FEI-6189] Disallow deploying frontend and backend changes at the same time. (#297)
## Summary:
We want to disallow deploying frontend and backend changes together (in general and in preperation for moving the frontend code out into a new repo).

We're not going to deploy this until we're absolutely sure that we've figured out any edge cases, but it's likely that we're really close to deploying this.

All the changes we've made are in these two epics:
https://khanacademy.atlassian.net/browse/FEI-6194
https://khanacademy.atlassian.net/browse/FEI-6178

Issue: FEI-6189

## Test plan:
Not sure how to test this... I think this is syntactically correct, but I assume the test plan is probably deploying this and then confirming that a plain static (or non-static) deploy continues to work - and that a combination deploy correctly fails.

Author: jeresig

Reviewers: jwbron, csilvers, somewhatabstract

Required Reviewers:

Approved By: jwbron, csilvers, somewhatabstract

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/297