[🔥AUDIT🔥] Enable the ability to deploy non master branches to buildmaster (#337)
🖍 _This is an audit!_ 🖍

## Summary:
We are working on splitting up the FE code out of the webapp repo and implement its own deploy system.  And we need to be able to deploy to buildmaster from a branch that is not master!  So we are making a change in jenkins to enable that temporarily and we will revert this as soon as we are done this weekend.

Issue: https://khanacademy.slack.com/archives/C01120CNCS0/p1749233546013349

## Test plan:
we are going to deploy these changes and retry deploying to buildmaster

Author: MiguelCastillo

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/337