Use `input` instead of a deploy param, to control deploying to production.
build-and-deploy-publish-worker.groovy has an input parameter to say
whether it does just-build, or build-and-deploy.  This PR changes it
to use an interactive prompt instead, just like we do with
deploy-fastly.groovy.  That way, a user can look at the output of the
build step and decide if it looks good or not, and then say whether to
deploy or not based on that.

Issue: https://github.com/Khan/jenkins-jobs/pull/323#discussion_r2089652443

Test plan:
Fingers crossed