[infra-10711-remove-static-deploys] Remove static deploys
The static deploy scripts have already been removed from webapp so
remove the dead code around them.

jobs/demo-district-update-pass should be fixed (likely moved to a GitHub
action in the frontend repo) in a separate PR before it's scheduled to
run again on August 1st.

Issue: https://khanacademy.atlassian.net/browse/INFRA-10711

Test plan:

build-webapp and deploy-znd can be tested with the replay feature.

deploy-webapp will need to be monitored and tested during a real
deployment in #1s-and-0s-deploys. It's too integrated with buildmaster
to be tested independently. I will be blocking the deploy queue during
this time.