[infra-10684-merge-branches-failure-better-slack] Propagate merge-branches errors to slack.
Include failure reason and stack trace when sending failure
notifications to slack.

For merge-branches, include the output of the `git checkout` and `git
merge` commands.

This is the start of an ongoing effort to push the reason for Jenkins
job failures to the deployer via slack. Right now, we direct users to
"look for the red circle(s)" in Jenkins which is fraught with unrelated
information that look like errors to anyone without intimate knowledge
of the deploy system. Ideally, users would never need to go to Jenkins
to know why their deployment failed and what actions they need to take.

Some context on the motivations for doing so can be found in this PR:

https://github.com/Khan/jenkins-jobs/pull/347

With notify.fail() now propagating the reason for failure, we can begin
to instrument the rest of our jobs in the same manner.