Make buildmaster update message more clear (#347)
## Summary:
Some folks have been getting confused by the following message in the
merge-branches job logs:

```
17:22:43  Marking commit #12345 as failed: null
```

This leads users to think this is an error message with a null reason as
opposed to what it actually is: a log of the status update we're sending
to buildmaster.

This happens whenever merge-branches encounters a failure, such as a
merge conflict or, in the reported case, when an commit-ish was passed
that doesn't exist in origin. The actual error message exists at the
bottom of the console logs and in the pipeline steps view next to the
red (X).

Change this message to:

```
17:53:27  Marking buildmaster commit #12345 as failed: No merged SHA
```

Issue: https://khanacademy.atlassian.net/browse/INFRA-10684

Test plan:

Tested with the replay feature:

https://jenkins.khanacademy.org/job/deploy/job/merge-branches/324339/console

Author: nathanjd

Reviewers: MiguelCastillo, csilvers, jwbron

Required Reviewers:

Approved By: MiguelCastillo

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/347