Post correct Google Logs urls during build/deploy (#360)
## Summary:

This is the second attempt at fixing Google Console Logs Explorer urls that jenkins posts to Slack during builds/deploys. 

The first time I deployed this it caused failures during deploys because of an obscure error causing Java not to be able to find the Groovy function `logViewerUrl` (which is the function I created to construct these logs urls).

It turns out that if the caller passes values that don't match the parameter list types, you just get a method-not-found error. 

With @nathanjd's help (well, let's be clear, I watched and Nathan figured it out) we found out the root problem and fixed it (which is to leave the `services` parameter as a `def` instead of trying to strongly type it).

Issue:

## Test plan:

Author: jeremywiebe

Reviewers: jeremywiebe, nathanjd, csilvers

Required Reviewers:

Approved By: nathanjd, csilvers

Checks:

Pull Request URL: https://github.com/Khan/jenkins-jobs/pull/360