Rename 'modules' to 'services' everywhere.
We are getting rid of the concept of 'modules', which are
appengine-specific.  In cloud-run-land they are just called
'services', and that's what we're going to call them too.

The main change is to use the new `--services` flag to
delete_gae_versions.py, which I just deployed today.  (A future change
will rename that file to something more appropriate.)

Issue: https://khanacademy.atlassian.net/browse/INFRA-10869

Test plan:
Fingers crossed